requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-filter==23.2
django-redis==5.4.0
django-cors-headers==4.1.0
psycopg2-binary==2.9.6
gunicorn==20.1.0
celery==5.3.1
redis==4.5.5
Pillow==9.5.0
python-dotenv==1.0.0